LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/mathml/presentation-markup/fractions/frac-parameters-2-expected.txt


PASS BottomDisplayStyleShiftDown 
PASS BottomDisplayStyleShiftDown Displaystyle 
FAIL BottomDisplayStyleShiftDown Displaystyle om fraction assert_approx_equals: mfrac: denominator shift expected 50 +/- 1 but got 0
PASS BottomShiftDown 
PASS DisplayStyleGapMin 
PASS DisplayStyleGapMin Dsiplaystyle 
FAIL DisplayStyleGapMin Dsiplaystyle on fraction assert_approx_equals: mfrac: gap expected 40 +/- 1 but got 0
PASS GapMin 
PASS TopDisplayStyleShiftUp 
PASS TopDisplayStyleShiftUp Displaystyle 
FAIL TopDisplayStyleShiftUp Displaystyle on fraction assert_approx_equals: mfrac: numerator shift expected 30 +/- 1 but got 0
PASS ToShiftUp 
FAIL ToShiftUp with Displaystyle on fraction assert_approx_equals: mfrac: numerator shift expected 90 +/- 1 but got 0
